refactor: change CoreText.kt package to composefix
