+++ app/src/main/java/app/meatin/ui/composables/CommentItem.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import app.meatin.ui.theme.CoreText
+import app.meatin.ui.theme.composefix.CoreText
 import app.meatin.ui.theme.MeatInTypography
 import coil.annotation.ExperimentalCoilApi
 import coil.compose.rememberImagePainter
