+++ app/src/main/java/app/meatin/ui/composables/PostCard.kt
@@ -31,7 +31,7 @@ import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import app.meatin.ui.theme.CoreText
+import app.meatin.ui.theme.composefix.CoreText
 import app.meatin.ui.theme.DarkFlamingo
 import app.meatin.ui.theme.MeatInTypography
 import coil.annotation.ExperimentalCoilApi
